ci(github-actions): add workflow to build and publish container images
This commit introduces a new GitHub Actions workflow to automate the building and publishing of Docker container images for both the `myncer-server` and `myncer-web` services.

The workflow is triggered on every push to the `master` branch, tagging the resulting images as `latest`. It also runs on new releases, tagging the images with the corresponding version number.